Implement no more movie to load behavior and clean up image downloading.
Prevent network calls and display a message either when there are no results for a search query or, if a search returned movie reviews, when there are no more results to that search.

Remove a previously placed but unused error label from SearchController and associated unit tests.

Implement a new label and showing/hiding behavior in the LoadingViewController to be inherited by collection views that inherit from it.

Remove the hasMore field from the MovieResult model since it was unused.

Update the result of network calls from SearchController and CriticsPickController to either return movies or relay there is more content to load. Similarly, update search bar delegate methods to update when there are no more reviews to load.

Add and layout the noResults message in the footer supplementary view of the collection view. Add hasMoreToLoad to the guard statement with the check that it’s not already loading before making another network call.

Implement onReuse and network request cancellation to clean up image loading. Remove image view delegate property, downloadImage method and all image view delegate conformance and usage. Add onReuse property, logic and downloadImage method to both collection view cells.