Reimplement and refactor movie image placeholder behavior to stop placeholder stretching.
Remove existing placeholder image from Assets. Replace with a programmatic SFSymbol in the Constants file.

Remove custom image view because it is no longer performing custom image download behavior. Replace with a static property in a UIImage extension to configure the movie image including setting the placeholder as the image view’s image.

Update the MovieCell protocol to replace the custom image view with a regular image view. Initialize the new image view in both collection view cells.

As placeholder image now set on the image view directly at creation, remove error paths to set the image to the placeholder from the image fetching function. in prepareForReuse, reset the image view image to the placeholder rather than nil.

Specify aspect ratio to .aspectFill when assigning a movie image and reset to .center in prepareForReuse so that the placeholder SFSymbol is centered and doesn’t stretch with the size of the image view (which changes in the SearchResultCell based on the height of text content in the cell).